Add Powershell equivalent of `tail -f`
Powershell equivalent of Unix `tail --follow` or better known as
`tail -f`.

This alias is born of the fact that this very handy command for
most Unix users that are watching logs. This capability is certain
to be useful to MS Windows develpers too.

The underlying `Get-Content` command is fairly easy, yes. However,
now, one does not need to look it up.

   --- Lionel Saliou, Ph.D